+++ app/src/main/java/com/codebroth/rewake/reminder/data/ReminderRepository.kt
@@ -7,7 +7,7 @@ interface ReminderRepository {
 
     fun getAllReminders(): Flow<List<Reminder>>
 
-    suspend fun insertReminder(reminder: Reminder): Long
+    suspend fun insertReminder(reminder: Reminder): Int
 
     suspend fun updateReminder(reminder: Reminder)
